+++ code/src/cljs/sixsq/nuvla/ui/edges/views.cljs
@@ -18,6 +18,7 @@
     [sixsq.nuvla.ui.panel :as panel]
     [sixsq.nuvla.ui.plugins.full-text-search :as full-text-search-plugin]
     [sixsq.nuvla.ui.plugins.pagination :as pagination-plugin]
+    [sixsq.nuvla.ui.session.subs :as session-subs]
     [sixsq.nuvla.ui.utils.forms :as utils-forms]
     [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.map :as map]
@@ -732,10 +733,11 @@
 
 
 (defn NuvlaboxRow
-  [{:keys [id name description created state tags online refresh-interval] :as _nuvlabox} managers]
+  [{:keys [id name description created state tags online refresh-interval created-by] :as _nuvlabox} managers]
   (let [uuid                  (general-utils/id->uuid id)
         locale                (subscribe [::i18n-subs/locale])
-        next-heartbeat-moment @(subscribe [::subs/next-heartbeat-moment id])]
+        next-heartbeat-moment @(subscribe [::subs/next-heartbeat-moment id])
+        creator               (subscribe [::session-subs/resolve-user created-by])]
     [ui/TableRow {:on-click #(dispatch [::history-events/navigate (str "edges/" uuid)])
                   :style    {:cursor "pointer"}}
      [ui/TableCell {:collapsing true}
@@ -745,6 +747,7 @@
      [ui/TableCell (or name uuid)]
      [ui/TableCell description]
      [ui/TableCell (values/format-created created)]
+     [ui/TableCell @creator]
      [ui/TableCell (when next-heartbeat-moment (utils/last-time-online next-heartbeat-moment refresh-interval @locale))]
      [ui/TableCell [uix/Tags tags]]
      [ui/TableCell {:collapsing true}
@@ -792,6 +795,7 @@
         [ui/TableHeaderCell "name"]
         [ui/TableHeaderCell "description"]
         [ui/TableHeaderCell (@tr [:created])]
+        [ui/TableHeaderCell (@tr [:created-by])]
         [ui/TableHeaderCell (@tr [:last-online])]
         [ui/TableHeaderCell "tags"]
         [ui/TableHeaderCell "manager"]]]